Support (multiple) inherited branches
Using a commit message prefix like [child:parent1:parent2] will make
git-branchless base child on parent1 and parent2.

You must ensure that dependencies appear before users. This is not enforced
at the moment, so with wrong commit order, git-branchless will silently base
child off an old version of parent1 and parent2.

Also this change will likely break people using a custom branchless.subjectRegex.
Previously the contract was that the first capture group is the branch name,
and the second one is the commit subject.
If we need more flexibility I'll redesign this anyway.