Preserve original commit timestamp when rebasing
The goal here is to make the commit ID of the rebased commit consistent
every time (both between topics and between `git-branchstack` runs),
which combined with `--trim-subject` enables workflows like building
chains of GitHub pull requests.

In a perfect world maybe we'd use the current user's name/email in the
committer field and just keep the timestamp from the original commit,
but it doesn't seem like there's a simple way to do that with
`git-revise` right now. In practice I assume the author and committer
will usually be the same person with this workflow anyway.

This allows us to correct the check if a branch should be updated.
Instead of comparing messages and trees, just compare OIDs.