Make --trim-subject the default and add --keep-tags to not trim subjects
With deterministic commit IDs we can now reuse dependencies' commit
objects.  For example if A depends on B, we now create A as descendant
of B.

This means that remotes/tools that see both A and B can exploit that
they are based of each other.

For example, as noted by Ryan Fitzgerald, if a GitHub pull request
targets its dependency, only changes on top of the dependency are
shown. When the dependency is merged, the pull requested will be
retargeted automatically.